[HMTM-187] - Logic update
