+++ src/profile/routes.py
@@ -4,15 +4,14 @@ from fastapi.requests import Request
 from fastapi.responses import HTMLResponse, RedirectResponse, Response
 
 from src.cookies import set_cookie
-from src.sso import get_me as get_me_dependency
+from src.common.dependencies import get_me as get_me_dependency
 from src.profile.dependencies import (
     change_password as change_password_dependency,
     update_user_profile as update_user_profile_dependency,
     master_by_user as master_by_user_dependency,
     update_master as update_master_info_dependency,
-    register_master as register_master_dependency,
+    register_master as register_master_dependency
 )
-from src.common.dependencies import active_user_session as check_active_user_session_dependency
 from src.sso.dto import (
     GetMeResponse
 )
@@ -31,18 +30,13 @@ from src.utils import (
 )
 
 templates = Jinja2Templates(directory="templates")
-router = APIRouter(
-    prefix="/profile",
-    tags=["Profile"],
-    dependencies=[Depends(check_active_user_session_dependency)]
-)
+router = APIRouter(prefix="/profile", tags=["Profile"])
 
 
-@router.get("/me", response_class=HTMLResponse, name="profile", dependencies=[])
+@router.get("/me", response_class=HTMLResponse, name="profile")
 async def profile_page(
         request: Request,
-        current_user: GetMeResponse = Depends(get_me_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
+        current_user: GetMeResponse = Depends(get_me_dependency)
 ):
     if current_user.user is not None:
         master: GetUserIsMasterResponse = await master_by_user_dependency(
@@ -68,16 +62,20 @@ async def profile_page(
 
         return response
 
-    encrypted_error = encryptor.encrypt(str(current_user.error))
-    return RedirectResponse(url=f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
+    return RedirectResponse(url="/sso/login", status_code=status.HTTP_303_SEE_OTHER)
 
 
 @router.post("/change-password", response_class=HTMLResponse, name="change-password")
 async def process_change_password(
+        current_user: GetMeResponse = Depends(get_me_dependency),
         result: ChangePasswordResponse = Depends(change_password_dependency),
         encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
 ):
     """ Ручка авторизованного пользователя для смены пароля - Процесс """
+    if current_user.user is None:  # Нужен для Redirect на login-page, т.к ошибка из result - ошибка конкретной логики
+        encrypted_error: str = encryptor.encrypt(str(current_user.error))
+        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
+
     if result.error is not None:
         encrypted_error_key: str = encryptor.encrypt(str(result.error))
 
@@ -100,10 +98,17 @@ async def process_change_password(
 
 @router.post("/update_user_profile", response_class=HTMLResponse, name="update_user_profile")
 async def process_update_user_profile(
+        request: Request,
         result: UpdateUserProfileResponse = Depends(update_user_profile_dependency),
         encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
 ):
     """ Ручка авторизованного пользователя для изменения данных о себе: никнейм, телеграм, телефон - Процесс """
+    current_user: GetMeResponse = await get_me_dependency(request=request, cookies=result.cookies)
+
+    if current_user.user is None:  # Нужен для Redirect на login-page, т.к ошибка из result - ошибка конкретной логики
+        encrypted_error: str = encryptor.encrypt(str(current_user.error))
+        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
+
     if result.error is not None:
         encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
 
@@ -126,9 +131,14 @@ async def process_update_user_profile(
 
 @router.post("/update-master", response_class=HTMLResponse, name="update-master")
 async def process_update_master(
+        current_user: GetMeResponse = Depends(get_me_dependency),
         result: UpdateMasterResponse = Depends(update_master_info_dependency),
         encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
 ):
+    if current_user.user is None:  # Нужен для Redirect на login-page, т.к ошибка из result - ошибка конкретной логики
+        encrypted_error: str = encryptor.encrypt(str(current_user.error))
+        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
+
     if result.error is not None:
         encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
 
@@ -151,9 +161,14 @@ async def process_update_master(
 
 @router.post("/register-master", response_class=HTMLResponse, name="register-master")
 async def register_master(
+        current_user: GetMeResponse = Depends(get_me_dependency),
         result: RegisterMasterResponse = Depends(register_master_dependency),
         encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
 ):
+    if current_user.user is None:  # Нужен для Redirect на login-page, т.к ошибка из result - ошибка конкретной логики
+        encrypted_error: str = encryptor.encrypt(str(current_user.error))
+        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
+
     if result.error is not None:
         encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
 
@@ -162,7 +177,8 @@ async def register_master(
             status_code=status.HTTP_303_SEE_OTHER
         )
 
-    encrypted_message_key: str = encryptor.encrypt("Вы успешно стали мастером!")
+    encrypted_message: FernetEnvironmentsKey = FernetEnvironmentsKey()
+    encrypted_message_key: str = encrypted_message.encrypt("Вы успешно стали мастером!")
 
     response: Response = RedirectResponse(
         url=f"/profile/me?message={encrypted_message_key}&tab=master",
