--- src/sso/__init__.py
@@ -1,1 +1,0 @@
-from src.sso.dependencies import get_me
