+++ src/sso/dependencies.py
@@ -27,7 +27,6 @@ from graphql_client import (
     ChangeForgetPasswordMutation,
     GetUserByIDQuery,
     GetUserByEmailQuery,
-    GetMeQuery,
     GetMasterByUserQuery,
 
     extract_error_message,
@@ -36,7 +35,6 @@ from graphql_client import (
 )
 from src.sso.dto import (
     LoginResponse,
-    GetMeResponse,
     RegisterResponse,
     VerifyEmailResponse,
     SendVerifyEmailMessageResponse,
@@ -49,7 +47,6 @@ from src.sso.models import (
     UserInfo,
 )
 from src.profile.models import Master
-from src.sso.utils import user_from_dict
 
 
 async def process_register(  # type: ignore[return]
@@ -189,73 +186,6 @@ async def refresh_tokens(
         )
 
         result.error = error
-
-    return result
-
-
-async def get_me(
-        request: Request,
-        cookies: Optional[List[CookiesConfig]] = None
-) -> GetMeResponse:
-    result: GetMeResponse = GetMeResponse()
-
-    if not cookies and len(request.cookies) == 0:
-        result.error = "Пользователь не найден"
-        return result
-
-    actual_cookies: Dict[str, str] = request.cookies
-    if cookies:
-        for cookie in cookies:
-            actual_cookies[cookie.KEY] = cookie.VALUE
-
-    try:
-        if request.cookies:
-            gql_response: GQLResponse = await config.graphql_client.gql_query(
-                query=GetMeQuery.to_gql(),
-                variable_values={},
-                cookies=actual_cookies
-            )
-            if "errors" in gql_response.result:
-                raise Exception
-
-            result.user = user_from_dict(gql_response.result["me"])
-
-    except Exception:
-        try:
-            refreshed_tokens: RefreshTokensResponse = await refresh_tokens(request=request, cookies=cookies)
-
-            if refreshed_tokens.error is not None:
-                result.error = refreshed_tokens.error
-                return result
-
-            result.headers = refreshed_tokens.headers
-            result.cookies = refreshed_tokens.cookies
-
-            actual_cookies = {}
-            for cookie in refreshed_tokens.cookies:
-                actual_cookies[cookie.KEY] = cookie.VALUE
-
-            gql_get_me: GQLResponse = await config.graphql_client.gql_query(
-                query=GetMeQuery.to_gql(),
-                variable_values={},
-                cookies=actual_cookies
-            )
-
-            if "errors" in gql_get_me.result:
-                result.error = gql_get_me.result["errors"][0]["message"]
-                return result
-
-            result.user = user_from_dict(gql_get_me.result["me"])
-
-        except Exception as err:
-            error = ERRORS_MAPPING.get(
-                extract_error_message(
-                    error=str(err),
-                    default_message="Ошибка проверки cookies"
-                ),
-                DEFAULT_ERROR_MESSAGE
-            )
-            result.error = error
 
     return result
